requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-dotenv
pydantic
bcrypt
PyJWT
